requirements.txt
customtkinter==5.2.1
esptool==4.1
mttkinter==0.6.1
paho_mqtt==1.6.1
pyserial==3.5
Requests==2.31.0
packaging==23.2
pyinstaller==5.10.1
argparse==1.4.0